Add files via upload
total services created

+++ lms/src/main/java/edu/unimagdalena/lms/services/StudentServiceImpl.java
@@ -4,24 +4,39 @@ import edu.unimagdalena.lms.entities.Student;
 import edu.unimagdalena.lms.repositories.StudentRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class StudentServiceImpl implements StudentService {
     private final StudentRepository studentRepository;
+
     public StudentServiceImpl(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
     }
 
     @Override
     public Student save(Student student) {
+        if(student == null){
+            throw new IllegalArgumentException("El estudiante no puede ser nulo.");
+        }
+        if (student.getEmail() == null || student.getEmail().isEmpty()) {
+            throw new IllegalArgumentException("El email es obligatorio");
+        }
+        if (student.getFullName() == null || student.getFullName().isEmpty()){
+            throw new IllegalArgumentException("El nombre es obligatorio.");
+        }
+        if(student.getCreatedAt() == null){
+            student.setCreatedAt(Instant.now());
+        }
+        student.setUpdatedAt(Instant.now());
         return studentRepository.save(student);
     }
 
     @Override
-    public Optional<Student> findById(Long id) {
-        return studentRepository.findById(id);
+    public Student findById(Long id) {
+        return studentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Estudiante no encontrado"));
     }
     @Override
     public List<Student> findAll() {
@@ -29,6 +44,9 @@ public class StudentServiceImpl implements StudentService {
     }
     @Override
     public void deleteById(Long id) {
+        if(!studentRepository.existsById(id)){
+            throw new RuntimeException("No se puede eliminar, estudiante no existe");
+        }
         studentRepository.deleteById(id);
     }
 }
